+++ Classes/ViewHelpers/PageViewHelper.php
@@ -3,6 +3,8 @@ namespace T3v\T3vDataMapper\ViewHelpers;
 
 use \T3v\T3vCore\ViewHelpers\AbstractViewHelper;
 
+use \T3v\T3vDataMapper\Service\PageService;
+
 /**
  * Page View Helper Class
  *
@@ -10,6 +12,8 @@ use \T3v\T3vCore\ViewHelpers\AbstractViewHelper;
  */
 class PageViewHelper extends AbstractViewHelper {
   /**
+   * The page service.
+   *
    * @var \T3v\T3vDataMapper\Service\PageService
    * @inject
    */
@@ -19,13 +23,13 @@ class PageViewHelper extends AbstractViewHelper {
    * The View Helper render function.
    *
    * @param int $uid The UID of the page
-   * @param boolean $languageOverlay If set, the language record (overlay) will be applied, defaults to `true`
+   * @param boolean $languageOverlay If set, the language record (overlay) will be applied
    * @param int $sysLanguageUid The optional system language UID, defaults to the current system language UID
    * @return object The page object
    */
-  public function render($uid, $languageOverlay = true, $sysLanguageUid = null) {
+  public function render($uid, $languageOverlay = null, $sysLanguageUid = null) {
     $uid             = intval($uid);
-    $languageOverlay = (boolean) $languageOverlay;
+    $languageOverlay = (boolean) $languageOverlay ?: null;
     $sysLanguageUid  = intval($sysLanguageUid) ?: $this->languageService->getSysLanguageUid();
 
     return $this->pageService->getPageByUid($uid, $languageOverlay, $sysLanguageUid);