+++ Configuration/TypoScript/Extensions/T3vDataMapper/constants.ts
@@ -19,7 +19,7 @@ plugin {
     }
 
     settings {
-      # ...
+      languageOverlay = 1
     }
   }
 }